backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic>=2.0.0
langchain
langchain-openai
langgraph
langchain-core
